refactor: move webgl examples
