+++ examples/webgl-with-svg-overlay/README.md
@@ -1,0 +1,3 @@
+# WebGL With SVG Overlay
+N.B. This example must be run via a web server in order for the data file to
+be served, this can be done by running `npm start` from the project root.
