+++ examples/webgl-with-svg-overlay/index.html
@@ -1,0 +1,28 @@
+<html>
+
+<head>
+    <script src="../../node_modules/seedrandom/seedrandom.js"></script>
+    <script>Math.seedrandom('a22ebc7c488a3a47');</script>
+    <script src="../../node_modules/d3/dist/d3.js"></script>
+    <script src="../../packages/d3fc/build/d3fc.js"></script>
+    <style>
+        body {
+            background-color: #000107;
+            color: white;
+        }
+
+        d3fc-canvas {
+            height: 90vh;
+        }
+    </style>
+</head>
+
+<body>
+    <d3fc-canvas use-device-pixel-ratio set-webgl-viewport></d3fc-canvas>
+    <input type="checkbox" id="showLabels" checked>
+    <label for="showLabels">Show labels</label>
+
+    <script src="index.js"></script>
+</body>
+
+</html>
