--- packages/d3fc-webgl/examples/candlestick.html
@@ -1,13 +1,0 @@
-<html>
-<head>
-  <script src="../node_modules/d3/dist/d3.js"></script>
-  <script src="../../d3fc/build/d3fc.js"></script>
-</head>
-<body>
-  <div>
-    <div id="chart" style="height: 90vh;"></div>
-  </div>
-
-  <script src="candlestick.js"></script>
-</body>
-</html>
