--- packages/d3fc-webgl/examples/line.html
@@ -1,15 +1,0 @@
-<html>
-<head>
-    <script src="../node_modules/d3/dist/d3.js"></script>
-    <script src="../../d3fc-extent/build/d3fc-extent.js"></script>
-    <script src="../../d3fc-random-data/build/d3fc-random-data.js"></script>
-    <script src="../../d3fc-series/build/d3fc-series.js"></script>
-    <script src="../build/d3fc-webgl.js"></script>
-</head>
-<body id="body">
-    <canvas id="line-canvas"  width="500" height="250" style="position: absolute; top: 8px; left: 8px;"></canvas>
-    <canvas id="line-webgl"  width="500" height="250"></canvas>
-    <p id="fps">n/a</p>
-    <script src="line.js"></script>
-</body>
-</html>
