--- packages/d3fc-webgl/examples/webgl-overlay.html
@@ -1,13 +1,0 @@
-<html>
-<head>
-    <script src="../node_modules/d3/dist/d3.js"></script>
-    <script src="../../d3fc/build/d3fc.js" ></script>
-</head>
-<body style="background: #000017; color: white;">
-    <div id="chart" style="height: 90vh"></div>
-    <input type="checkbox" id="showLabels" checked>
-    <label for="showLabels">Show labels</label>
-
-    <script src="webgl-overlay.js"></script>
-</body>
-</html>
