--- packages/d3fc-webgl/examples/webgl-performance.html
@@ -1,23 +1,0 @@
-<html>
-<head>
-    <script src="../node_modules/d3/dist/d3.js"></script>
-    <script src="../../d3fc/build/d3fc.js" ></script>
-</head>
-<body>
-    <div id="chart" style="height: 90vh"></div>
-    <p id="fps" style="float: right; margin: 0 30px;"></p>
-    <div>
-        <input type="radio" name="renderType" id="webgl" checked="true"/>
-        <label for="webgl">WebGL</label>
-
-        <input type="radio" name="renderType" id="canvas"/>
-        <label for="canvas">Canvas</label>
-    </div>
-    <div>
-        <input type="number" id="numPoints" min="10000" max="500000" value="100000" step="10000"/>
-        <label for="numPoints">Number data points</label>
-    </div>
-
-    <script src="webgl-performance.js"></script>
-</body>
-</html>
